Added solutions for day 19 parts 1 and 2.
The solution for part 2 is tricky. At the beginning I tried a BFS
approach, but of course it turned out not feasable at all, even
reversing the problem (from target to e). I tried DFS then and it
worked on the first branch (sorting the replacements by length). It
worked but only by pure luck. This won't work as a general solution.

It seems that the general solution to this might be A* algorithm.
TODO: learn more about that.